Moved authentication checks to tutorial group service
